requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==5.0.1
Werkzeug==3.1.3
google-generativeai==0.8.5
PyPDF2==3.0.1
python-dotenv==1.1.0
